cached parser
